Train code split, Part 3: moving some of the model.py code into sampler.py (#9)
Moving sampler code into sampler.py. Also moving some functions I found out are for training into train.py.
